requirements.txt
numpy
Pillow
Airtest
pytesseract
jupyterlab
uiautomator2
imutils
findit
scikit-image
pyyaml
git+https://github.com/r-luo/ocrspace.git
pandas